requirements.txt
ollama
requests
PyPDF2
logging
python-dotenv